backend/requirements.txt
# Core FastAPI and Server
fastapi==0.110.1
uvicorn[standard]==0.25.0
python-dotenv>=1.0.1
pydantic>=2.6.4
python-multipart>=0.0.9

# Database
pymongo==4.5.0
motor==3.3.1

# Shopify Integration
ShopifyAPI>=12.0.0
graphql-core>=3.2.0
requests>=2.31.0
requests-oauthlib>=2.0.0

# Security & Authentication
cryptography>=42.0.8
pyjwt>=2.10.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.2.0

# Caching & Performance
redis>=5.0.0
hiredis>=2.3.0

# AWS & Cloud Services
boto3>=1.34.129

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# Utilities
typer>=0.9.0
tzdata>=2024.2

# Development & Testing
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0

# Monitoring & Logging
structlog>=24.1.0
sentry-sdk[fastapi]>=1.40.0
